jobapp_agent/pyproject.toml
[project]
name = "jobapp_agent"
version = "0.1.0"
description = "Jobapp_agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.130.0,<1.0.0",
    "embedchain[postgres]>=0.1.128",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
]

[project.scripts]
jobapp_agent = "jobapp_agent.main:run"
run_crew = "jobapp_agent.main:run"
train = "jobapp_agent.main:train"
replay = "jobapp_agent.main:replay"
test = "jobapp_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
